Rename PreLoadRecord on checks directory
